Fix axis angle fabsf usage.

+++ matrix/AxisAngle.hpp
@@ -78,7 +78,7 @@ public:
         AxisAngle &v = *this;
         Type ang = (Type)2.0f*acosf(q(0));
         Type mag = sinf(ang/2.0f);
-        if (fabs(ang) < 1e-10f) {
+        if (fabsf(ang) < 1e-10f) {
             v(0) = 0;
             v(1) = 0;
             v(2) = 0;
